Traefik assertions.
